add support for wasm/wasip1 build without mmap support
