+++ mmap_unix.go
@@ -1,5 +1,5 @@
-//go:build !windows && !appengine && !plan9
-// +build !windows,!appengine,!plan9
+//go:build !windows && !appengine && !plan9 && !js && !wasip1
+// +build !windows,!appengine,!plan9,!js,!wasip1
 
 package maxminddb
