+++ mmap_windows.go
@@ -1,3 +1,4 @@
+//go:build windows && !appengine
 // +build windows,!appengine
 
 package maxminddb
