+++ reader_mmap.go
@@ -1,5 +1,5 @@
-//go:build !appengine && !plan9
-// +build !appengine,!plan9
+//go:build !appengine && !plan9 && !js && !wasip1
+// +build !appengine,!plan9,!js,!wasip1
 
 package maxminddb
 
@@ -9,10 +9,10 @@ import (
 )
 
 // Open takes a string path to a MaxMind DB file and returns a Reader
-// structure or an error. The database file is opened using a memory map,
-// except on Google App Engine where mmap is not supported; there the database
-// is loaded into memory. Use the Close method on the Reader object to return
-// the resources to the system.
+// structure or an error. The database file is opened using a memory map
+// on supported platforms. On platforms without memory map support, such
+// as WebAssembly or Google App Engine, the database is loaded into memory.
+// Use the Close method on the Reader object to return the resources to the system.
 func Open(file string) (*Reader, error) {
 	mapFile, err := os.Open(file)
 	if err != nil {
@@ -51,9 +51,7 @@ func Open(file string) (*Reader, error) {
 	return reader, nil
 }
 
-// Close unmaps the database file from virtual memory and returns the
-// resources to the system. If called on a Reader opened using FromBytes
-// or Open on Google App Engine, this method does nothing.
+// Close returns the resources used by the database to the system.
 func (r *Reader) Close() error {
 	var err error
 	if r.hasMappedFile {
